backend/requirements.txt
Django
djongo
pymongo
djangorestframework
requests
django-environ
django-mongodb-backend
django-cors-headers